feat: add update to punishment subs
Feat: add update to punishment command sub commands. This commit also includes an update for the createModLog function with the new update property. The duration model was also update due to the need of duration itself, and not the endsAt date only. Autocomplete.ts filter was also added the update subcommand due to the need of punishment ids autocomplete.