feat: interactions separate files
Interactions are now defined in a separate folder and a separate file. To define the peperties and the options of a slash command or a context menu, you have to make a new file in the path src/interactions/**/*.ts and define the options as @readonly. (if you want to see the available options, import interactionOptions from typings and define your command as interactionOptions.)